requirements.txt
pillow==10.3.0
numpy==1.26.4
matplotlib==3.9.0
keras-facenet==0.3.2
opencv-python==4.10.0.82
scipy==1.13.1
tensorflow==2.16.1
pandas==2.2.2
flask==3.0.3
streamlit==1.35.0
streamlit-webrtc==0.47.7
seaborn==0.13.2
scikit-learn==1.5.0